requirements.txt
# only needed for inspecting the EmojiDick.pdf file
#fitz==0.0.1.dev2 
#Pillow==10.2.0
beautifulsoup4
# Mastodon.py==1.8.1 or higher
Mastodon.py
python-dotenv
Requests
pydantic
